snort3_demo: Refactor run_test.sh script
This commit refactors the snort3_demo repository so that it can be executed
somewhat more portably, and ideally feels a bit more comfortable to the user.

These changes make one big assumption, namely:

 * snort and libdaq are installed to the default PREFIX directory
   (/usr/local).

   With this assumption made, the user is no longer responsible for
   specifying the install location for "snorty", i.e., `argv[1]` will
   now specify the path of tests to execute.

   NOTICE: If necessary, I will provide an override in the form of an
	   environment variable to restore specifying an alternate
	   install PREFIX.

Now argv[1] specifices the path to the "top" directory to search for,
and run tests from. The following commands are now all valid:

 * ./run_tests.sh tests/ips_options

    Runs all tests under "tests/ips_options/*"

 * ./run_tests.sh tests/ips_options/so_and_soid

    Runs the singular test under "tests/ips_options/so_and_soid/"

 * ./run_tests.sh tests/ips_options/so_and_soid/test.bats

    Identical to the previous exmaple, runs the singular test under
    "tests/ips_options/so_and_soid/"

 * ./run_tests.sh /root/alternate-path/some-private/integration-tests/

    Runs all the tests under the user specified external directory.

The last major refactor to the script is the way which tests are
"searched and exected". Previously, the run_tests.sh used `find` to
create a singular long command line argument of "directories" containing
the "tests.bat" script, passing it as an argument to a loop defined in
the same file.

The script is now chagned to use `find -exec run_one`, which is
functionally equivalent, but is now future proofed from maxing out
command line arguments, or size of a command line in the given shell.

NOTICE: Acceptance of this PR will surely result in further refactorying
        to make the environment more portable still.